Time: 8 ms (67.54%), Space: 48.2 MB (20.9%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.java
@@ -7,7 +7,7 @@ class Solution {
         }
         while(start<end){
             int mid = start + (end - start)/2;
-            long p = 0L;
+            int p = 0;
             for(int i =0;i<piles.length;i++){
                 if(piles[i]<=mid){
                     p=p+1;
